libs/shared-auth/pyproject.toml
[project]
name = "shared-auth"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "PyJWT[crypto]>=2.8",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared_auth"]
